+++ tests/TestCase/Controller/Admin/SystemsControllerTest.php
@@ -1,5 +1,6 @@
 <?php
 declare(strict_types=1);
+
 /**
  * This file is part of me-cms.
  *
@@ -156,7 +157,7 @@ class SystemsControllerTest extends ControllerTestCase
             'webroot',
             'temporary',
             'plugins',
-            'phpExtensions',
+            'php',
             'elfinder',
             'cakephp',
             'cache',